Removed id from ColumnHeader
