+++ app/src/main/java/com/evrencoskun/tableviewsample/tableview/model/Cell.java
@@ -27,25 +27,19 @@ package com.evrencoskun.tableviewsample.tableview.model;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import com.evrencoskun.tableview.filter.IFilterableModel;
 import com.evrencoskun.tableview.sort.ISortableModel;
 
 /**
  * Created by evrencoskun on 11/06/2017.
  */
 
-public class Cell implements ISortableModel, IFilterableModel {
+public class Cell extends CellBase implements ISortableModel {
     @NonNull
     private final String mId;
-    @Nullable
-    private final Object mData;
-    @NonNull
-    private final String mFilterKeyword;
 
     public Cell(@NonNull String id, @Nullable Object data) {
+        super(data);
         this.mId = id;
-        this.mData = data;
-        this.mFilterKeyword = String.valueOf(data);
     }
 
     /**
@@ -58,24 +52,4 @@ public class Cell implements ISortableModel, IFilterableModel {
         return mId;
     }
 
-    /**
-     * This is necessary for sorting process.
-     * See ISortableModel
-     */
-    @Nullable
-    @Override
-    public Object getContent() {
-        return mData;
-    }
-
-    @Nullable
-    public Object getData() {
-        return mData;
-    }
-
-    @NonNull
-    @Override
-    public String getFilterableKeyword() {
-        return mFilterKeyword;
-    }
 }
